Resolved warnings (comparing signed with unsigned int, unused variable and alike).
* src/svncpp/client_property.cpp
  (Client::proplist, Client::propget, Client::revproplist): removed unused variable

* src/svncpp/path.cpp
  (Fixed_apr_temp_dir_get): changed type of "i" from int to size_t

* src/auth_dlg.cpp
  (AuthDlg::Data::Data): initialized labelUser and textUser

* src/diff_dlg.cpp
  (DiffDlg::Data::CheckControls, DiffDlg::Data::GetData): added "default:" to switch

* src/folder_browser.cpp
  (FolderBrowser::Data::FindClosestChild): removed unused variable "success"

* src/preferences.cpp
  (m_preferences): removed unused variable

* src/proportional_splitter.cpp
  (ProportionalSplitterWindow::OnSize): added explicit cast for conversion from
    "float" to "int"

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7044 65390229-12b7-0310-b90b-f21a5aa7ec8e

+++ src/proportional_splitter.cpp
@@ -97,7 +97,7 @@ ProportionalSplitterWindow::OnSize (wxSizeEvent & event)
     m_old_length_tot = new_length_tot;
 
     // Make the nescessary adjustments
-    SetSashPosition (new_length[0] - GetBorderSize ());
+    SetSashPosition ((int)new_length[0] - GetBorderSize ());
   }
 }
